Move cluster tests to separate module
